feat(gas-estimator): :sparkles: Add support for custom state overrides provided by the client
