+++ src/entrypoint/v0.6.0/utils.ts
@@ -1,18 +1,16 @@
-export function removeSpecialCharacters(revertReason: string): string {
-  const match = revertReason.match(/AA(\d+)\s(.+)/);
+export function getRequiredPrefundV6(userOp: {
+  paymasterAndData: string;
+  callGasLimit: bigint;
+  verificationGasLimit: bigint;
+  preVerificationGas: bigint;
+  maxFeePerGas: bigint;
+}) {
+  const multiplier = userOp.paymasterAndData !== "0x" ? 3n : 1n;
 
-  if (match) {
-    const errorCode = match[1]; // e.g., "25"
-    const errorMessage = match[2]; // e.g., "invalid account nonce"
-    const newMatch = `AA${errorCode} ${errorMessage}`.match(
-      // eslint-disable-next-line no-control-regex
-      /AA.*?(?=\\u|\u0000)/
-    );
-    if (newMatch) {
-      const extractedString = newMatch[0];
-      return extractedString;
-    }
-    return `AA${errorCode} ${errorMessage}`;
-  }
-  return revertReason;
+  const requiredGas =
+    userOp.callGasLimit +
+    userOp.verificationGasLimit * multiplier +
+    userOp.preVerificationGas;
+
+  return requiredGas * userOp.maxFeePerGas;
 }
