+++ src/entrypoint/v0.7.0/constants.ts
@@ -2,3 +2,6 @@ import { Address } from "viem";
 
 export const ENTRYPOINT_V7_ADDRESS: Address =
   "0x0000000071727De22E5E9d8BAf0edAc6f37da032";
+
+//https://github.com/eth-infinitism/account-abstraction/blob/6f02f5a28a20e804d0410b4b5b570dd4b076dcf9/contracts/core/EntryPoint.sol#L39
+export const INNER_GAS_OVERHEAD = 10000n;
