+++ src/gas-estimator/arbitrum/constants.ts
@@ -1,0 +1,2 @@
+export const NODE_INTERFACE_ARBITRUM_ADDRESS =
+  "0x00000000000000000000000000000000000000C8";
