+++ src/gas-estimator/mantle/constants.ts
@@ -1,0 +1,4 @@
+export const MANTLE_BVM_GAS_PRICE_ORACLE_ADDRESS =
+  "0x420000000000000000000000000000000000000F";
+
+export const MANTLE_L1_ROLL_UP_FEE_DIVISION_FACTOR = 1000000n;
