+++ src/gas-estimator/optimism/constants.ts
@@ -1,0 +1,2 @@
+export const OPTIMISM_L1_GAS_PRICE_ORACLE_ADDRESS =
+  "0x420000000000000000000000000000000000000F";
