+++ src/shared/utils.ts
@@ -1,37 +1,16 @@
-export function getRequiredPrefundV6(userOp: {
-  paymasterAndData: string;
-  callGasLimit: bigint;
-  verificationGasLimit: bigint;
-  preVerificationGas: bigint;
-  maxFeePerGas: bigint;
-}) {
-  const multiplier = userOp.paymasterAndData !== "0x" ? 3n : 1n;
+import { getRequiredPrefundV6 } from "../entrypoint/v0.6.0/utils";
+import { getRequiredPrefundV7 } from "../entrypoint/v0.7.0/utils";
+import {
+  isUserOperationV6,
+  UserOperation,
+} from "../gas-estimator/UserOperation";
 
-  const requiredGas =
-    userOp.callGasLimit +
-    userOp.verificationGasLimit * multiplier +
-    userOp.preVerificationGas;
+export function getRequiredPrefund(userOperation: UserOperation) {
+  if (isUserOperationV6(userOperation)) {
+    return getRequiredPrefundV6(userOperation);
+  }
 
-  return requiredGas * userOp.maxFeePerGas;
-}
-
-export function getRequiredPrefundV7(userOp: {
-  preVerificationGas: bigint;
-  verificationGasLimit: bigint;
-  callGasLimit: bigint;
-  paymasterVerificationGasLimit: bigint;
-  paymasterPostOpGasLimit: bigint;
-  maxFeePerGas: bigint;
-}) {
-  const requiredGas =
-    userOp.verificationGasLimit +
-    userOp.callGasLimit +
-    userOp.preVerificationGas +
-    userOp.paymasterVerificationGasLimit +
-    userOp.paymasterPostOpGasLimit +
-    userOp.preVerificationGas;
-
-  return requiredGas * userOp.maxFeePerGas;
+  return getRequiredPrefundV7(userOperation);
 }
 
 export function bumpBigIntPercent(big: bigint, percent: number) {
